# openapi/pawday-m2.2.bundle.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Pawday M2.3 Media and Search Foundation API",
    "version": "0.2.3",
    "description": "M2.1 identity, M2.2 delivery, M2.3 controlled media and derived search operations. PostgreSQL is authoritative; binary media is stored by a provider.",
    "license": {
      "name": "Private project"
    }
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
...
  "paths": {
    "/consumer/auth/phone/request-code": {
      "post": {
        "operationId": "post_consumer_auth_phone_request_code",
        "summary": "POST /consumer/auth/phone/request-code",
        "description": "LOGIN 可匿名，REVERIFY 绑定消费者会话及本人手机。200 表示 OTP 挑战、加密发送材料与 Outbox 同一事务提交，短信异步投递；MQ 停机不影响该提交。不返回验证码。过期/被替代挑战跳过发送。",
        "tags": [
          "consumer"
        ],
        "security": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReceiptEnvelope"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Invalid or missing credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Realm, permission or reverify denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Resource unavailable in actor scope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Version or idempotency conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "Dependency unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "x-implemented-in": "M2.1",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CodeRequest"
              }
            }
          }
        }
      }
    },
    "/consumer/auth/phone/verify": {
      "post": {
        "operationId": "post_consumer_auth_phone_verify",
        "summary": "POST /consumer/auth/phone/verify",
        "description": "M2.1 已实现。身份域由路径与服务端会话校验；请求禁止未知字段。",
        "tags": [
          "consumer"
        ],
        "security": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokensEnvelope"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Invalid or missing credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Realm, permission or reverify denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Resource unavailable in actor scope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Version or idempotency conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "Dependency unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "x-implemented-in": "M2.1",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PhoneLogin"
              }
            }
          }
        }
      }
    },
...
    "/admin/operations/outbox": {
      "get": {
        "operationId": "list_outbox",
        "summary": "list_outbox",
        "description": "最近 limit 条元数据视图；不返回事件 payload、手机号、验证码密文。历史分页后续实现。",
        "tags": [
          "admin"
        ],
        "security": [
          {
            "adminBearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OutboxList"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Invalid or missing credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Realm, permission or reverify denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Resource unavailable in actor scope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Version or idempotency conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "Dependency unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "x-implemented-in": "M2.2",
        "x-permission": "outbox.read",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          }
        ]
      }
    },
    "/admin/operations/outbox/stats": {
      "get": {
        "operationId": "outbox_stats",
        "summary": "outbox_stats",
        "description": "数据库权威 backlog、各状态数量、最老事件年龄、inbox DEAD 与短信 DEAD。Prometheus 指标另位于受 outbox.read 控制的 /actuator/prometheus。",
        "tags": [
          "admin"
        ],
        "security": [
          {
            "adminBearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OutboxStatsEnvelope"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Invalid or missing credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Realm, permission or reverify denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Resource unavailable in actor scope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Version or idempotency conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "Dependency unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "x-implemented-in": "M2.2",
        "x-permission": "outbox.read"
      }
    },
    "/admin/operations/outbox/{event_id}/replay": {
      "post": {
        "operationId": "replay_outbox",
        "summary": "replay_outbox",
        "description": "只重放 DEAD 原事件或失败的 DEAD_LETTER 转发记录；消费者业务已 PROCESSED 时仍去重。根事件 consumer/SMS DEAD 也可重放。校验 generation、幂等键、当前权限与 session/action=outbox.replay 二次验证；审计同事务；过期/被替代 OTP 禁止复活。",
        "tags": [
          "admin"
        ],
        "security": [
          {
            "adminBearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReplayReceiptEnvelope"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Invalid or missing credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Realm, permission or reverify denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Resource unavailable in actor scope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Version or idempotency conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "Dependency unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "x-implemented-in": "M2.2",
        "x-permission": "outbox.replay",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-Reverify-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z0-9_-]{43}$"
            },
            "description": "绑定当前 session/action、300 秒有效、成功写入时单次消费；事务失败回滚消费。"
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 16,
              "maxLength": 128
            },
            "description": "同主体及命令内唯一；同键不同载荷返回 409。"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReplayRequest"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Meta": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "request_id": {
            "type": "string",
            "format": "uuid"
          },
          "correlation_id": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "request_id",
          "correlation_id"
        ]
      },
      "Page": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "next_cursor": {
            "type": [
              "string",
              "null"
            ]
          },
          "has_more": {
            "type": "boolean"
          }
        },
        "required": [
          "next_cursor",
          "has_more"
        ]
      },
      "ErrorEnvelope": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "error": {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "retryable": {
                "type": "boolean"
              },
              "details": {
                "type": "object"
              }
            },
            "required": [
              "code",
              "message",
              "retryable",
              "details"
            ]
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "required": [
          "error",
          "meta"
        ]
      },
      "Tokens": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "access_token": {
            "type": "string",
            "pattern": "^[A-Za-z0-9_-]{43}$"
          },
          "refresh_token": {
            "type": "string",
            "pattern": "^[A-Za-z0-9_-]{43}$"
          },
          "expires_in": {
            "type": "integer",
            "minimum": 1
          },
          "session_id": {
            "type": "string",
            "format": "uuid"
          },
          "user_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          }
        },
        "required": [
          "access_token",
          "refresh_token",
          "expires_in",
          "session_id",
          "user_id"
        ]
      },
      "Receipt": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "version": {
            "type": "integer",
            "minimum": 0
          },
          "accepted_at": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "version",
          "accepted_at"
        ]
      },
...
      "CodeRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "phone_e164": {
            "type": "string",
            "pattern": "^\\+[1-9][0-9]{6,14}$"
          },
          "purpose": {
            "type": "string",
            "enum": [
              "LOGIN",
              "REVERIFY"
            ]
          }
        },
        "required": [
          "phone_e164",
          "purpose"
        ]
      },
      "PhoneLogin": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "phone_e164": {
            "type": "string",
            "pattern": "^\\+[1-9][0-9]{6,14}$"
          },
          "code": {
            "type": "string",
            "pattern": "^[0-9]{6}$"
          },
          "device_id": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128
          }
        },
        "required": [
          "phone_e164",
          "code",
          "device_id"
        ]
      },
...
      "TokensEnvelope": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Tokens"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
      "ReceiptEnvelope": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Receipt"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
...
      "OutboxRow": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "root_event_id": {
            "type": "string",
            "format": "uuid"
          },
          "event_type": {
            "type": "string"
          },
          "event_version": {
            "type": "integer",
            "minimum": 1
          },
          "transport_kind": {
            "type": "string",
            "enum": [
              "EVENT",
              "RETRY",
              "DEAD_LETTER"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "PUBLISHING",
              "PUBLISHED",
              "FAILED_RETRYABLE",
              "DEAD"
            ]
          },
          "attempt_count": {
            "type": "integer",
            "minimum": 0
          },
          "available_at": {
            "type": "string",
            "format": "date-time"
          },
          "lease_expires_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "last_error_code": {
            "type": [
              "string",
              "null"
            ]
          },
          "dead_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "generation": {
            "type": "integer",
            "minimum": 0
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "root_event_id",
          "event_type",
          "event_version",
          "transport_kind",
          "status",
          "attempt_count",
          "available_at",
          "lease_expires_at",
          "last_error_code",
          "dead_at",
          "generation",
          "created_at"
        ]
      },
      "OutboxList": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OutboxRow"
            }
          },
          "page": {
            "$ref": "#/components/schemas/Page"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "required": [
          "data",
          "page",
          "meta"
        ]
      },
      "OutboxStats": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "status_counts": {
            "type": "array",
            "items": {
              "type": "object",
              "additionalProperties": false,
              "properties": {
                "status": {
                  "type": "string",
                  "enum": [
                    "PENDING",
                    "PUBLISHING",
                    "PUBLISHED",
                    "FAILED_RETRYABLE",
                    "DEAD"
                  ]
                },
                "count": {
                  "type": "integer",
                  "minimum": 0
                }
              },
              "required": [
                "status",
                "count"
              ]
            }
          },
          "backlog": {
            "type": "integer",
            "minimum": 0
          },
          "oldest_event_age_seconds": {
            "type": "number",
            "minimum": 0
          },
          "inbox_dead": {
            "type": "integer",
            "minimum": 0
          },
          "sms_dead": {
            "type": "integer",
            "minimum": 0
          }
        },
        "required": [
          "status_counts",
          "backlog",
          "oldest_event_age_seconds",
          "inbox_dead",
          "sms_dead"
        ]
      },
      "OutboxStatsEnvelope": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/OutboxStats"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
      "ReplayRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "expected_generation": {
            "type": "integer",
            "minimum": 0
          },
          "reason_code": {
            "type": "string",
            "pattern": "^[A-Z0-9_]{4,80}$"
          }
        },
        "required": [
          "expected_generation",
          "reason_code"
        ]
      },
      "ReplayReceipt": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "event_id": {
            "type": "string",
            "format": "uuid"
          },
          "generation": {
            "type": "integer",
            "minimum": 1
          },
          "status": {
            "type": "string",
            "const": "PENDING"
          }
        },
        "required": [
          "event_id",
          "generation",
          "status"
        ]
      },
      "ReplayReceiptEnvelope": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ReplayReceipt"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
...
    },
    "securitySchemes": {
      "consumerBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "opaque256",
        "description": "consumer scoped opaque access token"
      },
      "merchantBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "opaque256",
        "description": "merchant scoped opaque access token"
      },
      "adminBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "opaque256",
        "description": "admin scoped opaque access token"
      }
    }
  }
}